+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0b] - 2024-07-11
+
+### Added
+
+- ASCII input format for voxelized geometries
+
+### Fixed
+
+- Materials creation via the configuration file fails for materials with a single element
+
 ## [1.11.0] - 2024-07-01
 
 ### Added